Restore multi-group trajectory execution
Following commit: 72cd3af7920fca0818375237fc375f7bab809445

This commit changes the default behavior of the Motoman driver when 
processing a `multi-group` trajectory with missing robot groups (i.g. 
When moving two arms but not the torso of the robot, or when moving only 
the torso and one of the arms):
- Currently the driver is zeroing out the trajectory positions for these 
motion groups.
- With the change added on this commit the driver will use the most 
recent feedback joint values as the default value instead.